finish wallpaper component, hover to elllipse button
