+++ src/pages/profiles/[id]/index.tsx
@@ -4,12 +4,10 @@ import {
 } from '@entities/profile';
 import { fetchUser, SWR_USER_KEY } from '@entities/user';
 import { DEFAULT_PUBLIC_KEY_STR } from '@shared/defaults';
-import EditIcon from '@shared/icons/Edit.svg';
-import { displayPublicKey, Avatar } from '@shared/ui';
-import { useConnection, useWallet } from '@solana/wallet-adapter-react';
+import { displayPublicKey, Avatar, Wallpaper } from '@shared/ui';
+import { useConnection } from '@solana/wallet-adapter-react';
 import { PublicKey } from '@solana/web3.js';
 import { NextPage } from 'next';
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 
@@ -19,11 +17,6 @@ const ProfilePage: NextPage = () => {
 
   const userId = (query.id as string) || DEFAULT_PUBLIC_KEY_STR;
   const userPK = new PublicKey(userId);
-
-  const wallet = useWallet();
-
-  const walletPublicKey = wallet.publicKey || PublicKey.default;
-  const isOwner = walletPublicKey.equals(userPK);
 
   const { data: userInfo, isLoading: isLoadingUserInfo } = useSWR(
     [SWR_USER_KEY, userId],
@@ -43,21 +36,8 @@ const ProfilePage: NextPage = () => {
   return (
     <div className="container grid grid-cols-8 gap-6 px-24">
       <div className="col-span-full mt-8 h-fit overflow-hidden rounded-xl bg-primary pb-8">
-        <div className="h-44 bg-[#D3EDFF]">
-          {isOwner && (
-            <div className="flex w-full justify-end p-4">
-              <Link
-                href={{
-                  pathname: '/profiles/[id]/edit',
-                  query,
-                }}
-              >
-                <button className="text-accent rounded-full bg-light-500 p-2 text-accent-500 shadow-md">
-                  <EditIcon />
-                </button>
-              </Link>
-            </div>
-          )}
+        <div className="h-44">
+          <Wallpaper />
         </div>
         <div className="flex flex-col px-6">
           <div className="-mt-16 mb-2">
