+++ src/shared/ui/ButtonEllipse.tsx
@@ -8,7 +8,7 @@ interface ButtonProps {
 const ButtonEllipse: React.FC<ButtonProps> = ({ children, onClick }) => {
   return (
     <button
-      className="text-accent rounded-full bg-light-500 p-2 text-accent-500 shadow-md"
+      className="text-accent rounded-full bg-light-500 p-2 text-accent-500 shadow-md transition hover:bg-accent-500 hover:text-light-500"
       onClick={onClick}
     >
       {children}
